Extract bullet form
