+++ src/pages/content/work-history/useWorkExperiencePanel.tsx
@@ -9,6 +9,7 @@ import WebsiteField from "../../../common/components/fields/WebsiteField";
 import SecondaryButton from "../../../common/components/SecondaryButton";
 import SlideOver from "../../../common/components/SlideOver";
 import { BulletPoint, WorkExperience } from "../../../common/interfaces/resume";
+import BulletForm, { FormBullet } from "../components/BulletForm";
 
 interface WorkExperienceForm {
   companyName: string;
@@ -56,10 +57,6 @@ function convertExperienceToFormData(
     endDateYear: experience.endDate?.year.toString(),
     isCurrentPosition: !experience.endDate,
   };
-}
-
-interface FormBullet extends BulletPoint {
-  shouldDelete: boolean;
 }
 
 /**
@@ -195,79 +192,7 @@ export default function useWorkExperiencePanel(
         </div>
       </div>
 
-      <div className="mt-12">
-        <h3 className="mb-4 text-base font-medium text-gray-700">
-          Bullet points
-        </h3>
-
-        <div className="flex flex-col gap-4">
-          {bullets.map((bullet, index) => (
-            <div key={bullet.id} className="flex items-center gap-2">
-              <textarea
-                className={classNames(
-                  "block w-full resize-none rounded-md border-gray-300 shadow-sm focus:border-gray-500 focus:ring-gray-500 sm:text-sm",
-                  {
-                    "cursor-not-allowed border-red-200 bg-red-50 text-red-400":
-                      bullet.shouldDelete,
-                  }
-                )}
-                rows={2}
-                disabled={bullet.shouldDelete}
-                value={bullet.text}
-                onChange={(e) => {
-                  const newBullets = bullets.slice();
-                  newBullets[index] = { ...bullet, text: e.target.value };
-                  setBullets(newBullets);
-                }}
-                autoFocus
-              />
-
-              <button
-                type="button"
-                className={classNames(
-                  "flex-shrink-0 inline-flex h-8 w-8 items-center justify-center rounded-full focus:outline-none focus:ring-2 focus:ring-gray-500 focus:ring-offset-2 no-mouse-focus-ring",
-                  {
-                    "text-gray-400 hover:text-gray-700": !bullet.shouldDelete,
-                    "text-red-400 hover:text-red-700": bullet.shouldDelete,
-                  }
-                )}
-                onClick={() => {
-                  const newBullets = bullets.slice();
-                  newBullets[index] = {
-                    ...bullet,
-                    shouldDelete: !bullet.shouldDelete,
-                  };
-                  setBullets(newBullets);
-                }}
-              >
-                <MinusCircleIcon className="h-5 w-5" aria-hidden="true" />
-              </button>
-            </div>
-          ))}
-
-          <div>
-            <SecondaryButton
-              onClick={() => {
-                setBullets([
-                  ...bullets,
-                  {
-                    id: generateId(),
-                    text: "",
-                    included: true,
-                    shouldDelete: false,
-                  },
-                ]);
-              }}
-            >
-              Add bullet point
-            </SecondaryButton>
-          </div>
-        </div>
-      </div>
+      <BulletForm bullets={bullets} onChange={setBullets} />
     </SlideOver>,
   ];
-}
-
-function generateId(): string {
-  return new Date().getTime().toString();
 }
